Data models update
* application renamed according to data model described in Miro app
* models reshuffled to appropriate applications
* whole new apps and models
