+++ backend/colleges/models.py
@@ -35,3 +35,29 @@ class Faculty(TimeStampedModel):
 
     def __str__(self):
         return f"{self.name} / {self.college} "
+
+
+class Course(TimeStampedModel):
+    """
+    Kurz.
+
+    Zdroj: https://docs.google.com/spreadsheets/d/1_karAzypSkiUOgrp6cm0_PLCimXyzdunxuUbdZKqjvI/edit#gid=0
+    """
+
+    name = models.CharField("Název", max_length=300)
+    faculty = models.ForeignKey(Faculty, on_delete=models.SET_NULL, null=True, verbose_name="Fakulta")
+    price = models.IntegerField("Cena")
+    sds = models.IntegerField("Standardní doba studia")
+    form_present = models.BooleanField("Prezenční forma studia", default=False)
+    form_combined = models.BooleanField("Kombinovaná forma studia", default=True)
+    url = models.URLField("URL na podrobnější informace")
+    subjects = models.ManyToManyField("teaching.Subject", verbose_name="Předměty")
+    note = models.TextField("Poznámka", blank=True)
+
+    class Meta:
+        verbose_name = "Kurz"
+        verbose_name_plural = "Kurzy"
+        ordering = ("name", "faculty", "faculty__college")
+
+    def __str__(self):
+        return f"{self.name} / {self.faculty}"
