--- backend/maps/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class MapsConfig(AppConfig):
-    name = "maps"
