+++ backend/qualifications/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import CollegeArea, CollegeProgramme
+from .models import CollegeArea, CollegeProgramme, EducationType, Title
 
 
 class CollegeAreaAdmin(admin.ModelAdmin):
@@ -14,5 +14,17 @@ class CollegeProgrammeAdmin(admin.ModelAdmin):
     filter_horizontal = ("subjects",)
 
 
+class EducationTypeAdmin(admin.ModelAdmin):
+    list_display = ("education", "qualification")
+    list_filter = ("qualification",)
+    search_fields = ("education",)
+
+
+class TitleAdmin(admin.ModelAdmin):
+    list_display = ("name", "shortcut")
+
+
 admin.site.register(CollegeArea, CollegeAreaAdmin)
 admin.site.register(CollegeProgramme, CollegeProgrammeAdmin)
+admin.site.register(EducationType, EducationTypeAdmin)
+admin.site.register(Title, TitleAdmin)
