--- backend/rvp/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class RvpConfig(AppConfig):
-    name = "rvp"
