+++ backend/zacniucit/settings.py
@@ -39,8 +39,8 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "django_extensions",
-    "rvp",
-    "maps",
+    "teaching",
+    "qualifications",
     "colleges",
 ]
